query_apply: nm: don't do connection down to remove routes
We were doing connection down & up when a route had to be removed
from an NM profile, instead of doing just a reapply. This was due to
some bugs in NM regarding removing routes:
- https://bugzilla.redhat.com/1837254
- https://issues.redhat.com/browse/RHEL-66262
- https://issues.redhat.com/browse/RHEL-67324

Those bugs are fixed since NM-1.51.5. However, instead of relying in the
version number to know if it's fixed, use the D-Bus VersionInfo
property, as NM is exposing there as a capability whether this bug is
fixed or not. This will facilitate to do the right thing on distros'
downstream builds if they fix the bug in NM via downstream patch.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>
(cherry picked from commit d6589384b159e85c54428af744a552f5340ae668)